requirements.txt
agno==1.1.13
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docstring-parser==0.16
duckduckgo-search==7.5.2
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
google-genai==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
primp==0.14.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
tomli==2.2.1
typer==0.15.2
typing-extensions==4.12.2
urllib3==2.3.0
websockets==14.2
